Base unit behaviour - all projectile dependencies move to Projectile folder
